% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_Gibbs_2d}
\alias{rcpp_Gibbs_2d}
\title{Folded Normal 2D RNG with Gibbs Sampling}
\usage{
rcpp_Gibbs_2d(N, x_init, sig1, sig2, rho)
}
\arguments{
\item{N}{sample size}

\item{x_init}{vector of length 2, initial value}

\item{sig1}{variance[1,1]}

\item{sig2}{variance[2,2]}

\item{rho}{rho}
}
\value{
matrix of N*2
}
\description{
Generate Random Number for 2-d Folded Normal with Gibss Sampling
}
