% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_Gibbs_nd}
\alias{rcpp_Gibbs_nd}
\title{Generate Multivariate (p>2) Gibbs Sampler}
\usage{
rcpp_Gibbs_nd(N, x_init, Sig)
}
\arguments{
\item{N}{Gibbs runs}

\item{x_init}{initial value of length p, p > 2}

\item{Sig}{}
}
\value{
Gibbs samples matrix of N * p
}
\description{
Generate Multivariate (p>2) Gibbs Sampler
}
