% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_symbolic_matrix}
\alias{rcpp_symbolic_matrix}
\title{Generate the Symbolic Matrix for Folded Normal}
\usage{
rcpp_symbolic_matrix(n)
}
\arguments{
\item{n}{dimensions, 1 <= n <= 16}
}
\value{
matrix of 2^n*n
}
\description{
Generate the Symbolic Matrix for Folded Normal
}
