\name{rfnorm-package}
\alias{rfnorm-package}
\alias{rfnorm}
\docType{package}
\title{\packageTitle{rfnorm}}
\description{\packageDescription{rfnorm}}
\section{Package Content}{\packageIndices{rfnorm}}
\author{\packageAuthor{rfnorm}}
\section{Maintainer}{\packageMaintainer{rfnorm}}
\keyword{package}
