% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfnorm.R
\name{mvrfnorm}
\alias{mvrfnorm}
\title{Multivriate Folded Normal Random Number Generation}
\usage{
mvrfnorm(
  n,
  mu,
  Sigma,
  method = "Gibbs",
  init = matrix(rep(0.5, length(mu) * 4), nrow = 4),
  gibbs_length = 1000,
  gibbs_chain = 4,
  checkSigma = FALSE
)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean vector}

\item{Sigma}{Variance matrix, should be postive, set checkSigma = TRUE to check the postiveness}

\item{method}{"Gibbs", "HMC", "mvtnorm", default = "Gibbs"}

\item{init}{initial value (matrix) for Gibbs sampler with p x chain_n dimesions. p is the variable dimension, chain_n is the number of chains.}

\item{gibbs_length}{iteration length for Gibbs sampler, default = 1000}

\item{gibbs_chain}{number of chains for Gibbs sampler, default = 4}

\item{checkSigma}{TRUE/FALSE If TRUE, then mvrfnorm checks whether the variance matrix is positive or not.}
}
\value{
sample array list
         - method "Gibbs" or "HMC" or "mvtnorm"
}
\description{
Generate multivriate Folded Normal random numbers
}
\examples{
ux = matrix(rnorm(4), 2, 2)
Sig = ux \%*\% t(ux)
sample_array = mvrfnorm(100,
c(0,0),
Sig,
init=matrix(c(
    0.5, 1.5,
    0.5, 1.5,
    1.5, 0.5,
    1.5, 1.5
    ), byrow = TRUE, ncol = 2), gibbs_chain =  4)
}
