% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_Gibbs_nd}
\alias{rcpp_Gibbs_nd}
\title{Generate the Symbolic Matrix for Folded Normal}
\usage{
rcpp_Gibbs_nd(N, x_init, Sig)
}
\arguments{
\item{N}{Gibbs runs}

\item{x_init}{initial value of length p}

\item{Sig}{}
}
\value{
Gibbs samples matrix of N * p
}
\description{
Generate the Symbolic Matrix for Folded Normal
}
